moved getSessionUser to api views
