+++ auth_backend/modules/user/views.py
@@ -6,13 +6,9 @@ Comments:
 from django.urls import reverse
 from django.views.generic import TemplateView
 from django.views.generic.edit import FormView
-from rest_framework import permissions
-from rest_framework.response import Response
-from rest_framework.views import APIView
 
 from auth_backend.modules.common.mixins import LoginRequiredMixin
 from auth_backend.modules.user.forms import ProfileForm
-from auth_backend.modules.user.serializers import UserSerializer
 
 
 class ProfileCreationView(FormView):
@@ -42,14 +38,3 @@ class ProfileSuccessfullyCreatedView(LoginRequiredMixin, TemplateView):
     Django view to redirect users to after they have successfully created their profile.
     """
     template_name = "user/profile_create_success.html"
-
-
-class GetRequestUserData(APIView):
-    """
-    API view to get a JSON representation of the current (session) user
-    """
-    permission_classes = (permissions.IsAuthenticated,)
-
-    def get(self, request):
-        serializer = UserSerializer(request.user)
-        return Response(serializer.data)
